# docs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/api/users/{id}/points/charge":{"post":{"tags":["user-controller"],"operationId":"chargePoint","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"integer","format":"int64"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/create":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/payments/{id}":{"post":{"tags":["payments-controller"],"operationId":"completePayment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"arg1","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/payments/{id}/cancel":{"post":{"tags":["payments-controller"],"operationId":"cancelPayment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"arg1","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/payments/create":{"post":{"tags":["payments-controller"],"operationId":"createPayment","parameters":[{"name":"arg0","in":"query","required":true,"schema":{"$ref":"#/components/schemas/PaymentRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/orders":{"post":{"tags":["order-controller"],"operationId":"orderGoods","parameters":[{"name":"arg0","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"arg1","in":"query","required":true,"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderRequest"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/orders/{id}/cancel":{"post":{"tags":["order-controller"],"operationId":"cancelOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"arg1","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/goods/create":{"post":{"tags":["goods-controller"],"operationId":"createGoods","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoodsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/coupons/{couponId}/get":{"post":{"tags":["coupon-controller"],"operationId":"getCoupon","parameters":[{"name":"couponId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"arg1","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/coupons/create":{"post":{"tags":["coupon-controller"],"operationId":"createCoupon","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCouponRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/{id}/points/check":{"get":{"tags":["user-controller"],"operationId":"checkPoint","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/{id}/coupons":{"get":{"tags":["user-controller"],"operationId":"checkCoupon","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/orders/{id}":{"get":{"tags":["order-controller"],"operationId":"getMyOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/goods":{"get":{"tags":["goods-controller"],"operationId":"findAllGoods","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GoodsResponse"}}}}}}}},"/api/goods/{id}":{"get":{"tags":["goods-controller"],"operationId":"getOneGoodsInfo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/goods/best":{"get":{"tags":["goods-controller"],"operationId":"getBestGoods","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SalesHistoryEntity"}}}}}}}},"/api/coupons":{"get":{"tags":["coupon-controller"],"operationId":"allCouponList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CouponEntity"}}}}}}}}},"components":{"schemas":{"UserRequest":{"type":"object","properties":{"name":{"type":"string"}}},"PaymentRequest":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"orderId":{"type":"integer","format":"int64"},"couponId":{"type":"integer","format":"int64"}}},"OrderRequest":{"type":"object","properties":{"goodsId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int64"}}},"GoodsRequest":{"type":"object","properties":{"goodsName":{"type":"string"},"price":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int64"}}},"CreateCouponRequest":{"type":"object","properties":{"couponName":{"type":"string"},"discountRate":{"type":"integer","format":"int64"},"capacity":{"type":"integer","format":"int64"},"dueDate":{"type":"string","format":"date"}}},"GoodsResponse":{"type":"object","properties":{"goodsId":{"type":"integer","format":"int64"},"goodsName":{"type":"string"},"price":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int64"}}},"SalesHistoryEntity":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"salesHistoryId":{"type":"integer","format":"int64"},"goodsId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int64"}}},"CouponEntity":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"couponId":{"type":"integer","format":"int64"},"couponName":{"type":"string"},"discountRate":{"type":"integer","format":"int64"},"capacity":{"type":"integer","format":"int64"},"dueDate":{"type":"string","format":"date"}}}}}}